feat: Display current user info in edit profile modal
