+++ src/app/[username]/@modal/(.)edit-profile/page.tsx
@@ -1,102 +1,47 @@
-"use client";
+import { EditProfileModal } from "@/components/profile/edit-profile-modal";
+import { getUserByUsername } from "@/lib/dal/users";
+import { auth } from "@clerk/nextjs/server";
+import { prisma } from "@/lib/prisma";
+import { notFound } from "next/navigation";
 
-import { useRouter } from "next/navigation";
-import { Button } from "@/components/ui/button";
-import { X } from "lucide-react";
-import { useEffect } from "react";
+interface PageProps {
+  params: { username: string };
+}
 
-export default function EditProfileModal() {
-  const router = useRouter();
+export default async function EditProfileInterceptingRoute({ params }: PageProps) {
+  const username = decodeURIComponent(params.username);
 
-  // ESCキーでモーダルを閉じる
-  useEffect(() => {
-    const handleEscape = (e: KeyboardEvent) => {
-      if (e.key === "Escape") {
-        router.back();
-      }
-    };
+  // 認証チェック
+  const { userId: clerkId } = await auth();
+  if (!clerkId) {
+    notFound();
+  }
 
-    document.addEventListener("keydown", handleEscape);
-    return () => document.removeEventListener("keydown", handleEscape);
-  }, [router]);
+  const currentUser = await prisma.user.findUnique({
+    where: { clerkId },
+    select: { id: true, username: true },
+  });
+
+  // 自分のプロフィールでない場合はアクセス拒否
+  if (!currentUser || currentUser.username !== username) {
+    notFound();
+  }
+
+  const user = await getUserByUsername(username);
+
+  if (!user) {
+    notFound();
+  }
 
   return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center">
-      {/* 背景オーバーレイ */}
-      <div
-        className="absolute inset-0 bg-black/50"
-        onClick={() => router.back()}
-      />
-
-      {/* モーダルコンテンツ */}
-      <div className="relative bg-white dark:bg-gray-900 rounded-2xl w-full max-w-2xl max-h-[90vh] overflow-y-auto shadow-xl m-4">
-        {/* ヘッダー */}
-        <div className="sticky top-0 bg-white dark:bg-gray-900 border-b border-gray-200 dark:border-gray-800 px-4 py-3 flex items-center justify-between">
-          <div className="flex items-center gap-8">
-            <button
-              onClick={() => router.back()}
-              className="rounded-full p-2 hover:bg-gray-100 dark:hover:bg-gray-800 transition"
-            >
-              <X className="h-5 w-5" />
-            </button>
-            <h2 className="text-xl font-bold">Edit Profile</h2>
-          </div>
-          <Button
-            size="sm"
-            className="bg-black text-white hover:bg-gray-800 dark:bg-white dark:text-black dark:hover:bg-gray-200 rounded-full px-4"
-          >
-            Save
-          </Button>
-        </div>
-
-        {/* コンテンツ */}
-        <div className="p-6">
-          {/* カバー画像 */}
-          <div className="mb-6">
-            <label className="block text-sm font-medium mb-2">Cover Image</label>
-            <div className="relative h-48 bg-gradient-to-r from-blue-400 to-purple-500 rounded-lg overflow-hidden">
-              <div className="absolute inset-0 bg-black/20"></div>
-              <div className="absolute inset-0 flex items-center justify-center">
-                <Button variant="secondary" size="sm">
-                  Change Cover
-                </Button>
-              </div>
-            </div>
-          </div>
-
-          {/* プロフィール画像 */}
-          <div className="mb-6">
-            <label className="block text-sm font-medium mb-2">Profile Image</label>
-            <div className="flex items-center gap-4">
-              <div className="w-24 h-24 rounded-full bg-gray-200 dark:bg-gray-700"></div>
-              <Button variant="secondary" size="sm">
-                Change Photo
-              </Button>
-            </div>
-          </div>
-
-          {/* 名前 */}
-          <div className="mb-6">
-            <label className="block text-sm font-medium mb-2">Name</label>
-            <input
-              type="text"
-              className="w-full px-4 py-2 border border-gray-300 dark:border-gray-700 rounded-lg bg-transparent focus:outline-none focus:ring-2 focus:ring-blue-500"
-              placeholder="Your name"
-            />
-          </div>
-
-          {/* Bio */}
-          <div className="mb-6">
-            <label className="block text-sm font-medium mb-2">Bio</label>
-            <textarea
-              rows={4}
-              className="w-full px-4 py-2 border border-gray-300 dark:border-gray-700 rounded-lg bg-transparent focus:outline-none focus:ring-2 focus:ring-blue-500 resize-none"
-              placeholder="Tell us about yourself"
-            />
-          </div>
-        </div>
-      </div>
-    </div>
+    <EditProfileModal
+      user={{
+        displayName: user.displayName,
+        bio: user.bio,
+        profileImageUrl: user.profileImageUrl,
+        coverImageUrl: user.coverImageUrl,
+      }}
+    />
   );
 }
